backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.4.0
django-filter==23.5
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
# JWT Authentication
djangorestframework-simplejwt==5.3.0
# Database - Use SQLite for now, can switch to PostgreSQL later
# psycopg2-binary==2.9.9
# API Documentation
drf-yasg==1.21.7
# Firebase
firebase-admin==6.4.0
# OpenAI/LLM Integration
openai==1.3.7
# Additional utilities
python-dateutil==2.8.2
# Celery and Redis (optional for now)
# celery==5.3.4
# redis==5.0.1
